Try restructuring as more standard Go package
