+++ cmd/adapter/mqttntrip/go.mod
