+++ cmd/adapter/mqttntrip/mqttntrip.go
