+++ cmd/adapter/ntripmqtt/go.mod
