+++ ntrip/mount.go
@@ -1,4 +1,4 @@
-package main
+package ntrip
 
 import (
 	"fmt"
