--- src/ntripgateway/server.go
@@ -1,1064 +1,0 @@
-package main
-
-import (
-	"context"
-	mqtt "github.com/eclipse/paho.mqtt.golang"
-	"github.com/google/uuid"
-	"github.com/gorilla/mux"
-	log "github.com/sirupsen/logrus"
-	"net/http"
-	"time"
-)
-
-func main() {
-	// TODO: Define from config file - would like to find config manager which is capable of invoking a goroutine for each element of list, as well as for new elements added to the list (watcher functions for mounts)
-	caster := &Caster{"2101", "go-ntrip.geops.team", "NTRIP Gateway for MQTT", "GA", "AUS", map[string]*Mount{
-		"TEST00AUS0": &Mount{Name: "TEST00AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS0": &Mount{Name: "TEST01AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS0": &Mount{Name: "TEST02AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS0": &Mount{Name: "TEST03AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS0": &Mount{Name: "TEST04AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS0": &Mount{Name: "TEST05AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS0": &Mount{Name: "TEST06AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS0": &Mount{Name: "TEST07AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS0": &Mount{Name: "TEST08AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS0": &Mount{Name: "TEST09AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS0": &Mount{Name: "TEST10AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS0": &Mount{Name: "TEST11AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS0": &Mount{Name: "TEST12AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS0": &Mount{Name: "TEST13AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS0": &Mount{Name: "TEST14AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS0": &Mount{Name: "TEST15AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS0": &Mount{Name: "TEST16AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS0": &Mount{Name: "TEST17AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS0": &Mount{Name: "TEST18AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS0": &Mount{Name: "TEST19AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS0": &Mount{Name: "TEST20AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS0": &Mount{Name: "TEST21AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS0": &Mount{Name: "TEST22AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS0": &Mount{Name: "TEST23AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS0": &Mount{Name: "TEST24AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS0": &Mount{Name: "TEST25AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS0": &Mount{Name: "TEST26AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS0": &Mount{Name: "TEST27AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS0": &Mount{Name: "TEST28AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS0": &Mount{Name: "TEST29AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS0": &Mount{Name: "TEST30AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS0": &Mount{Name: "TEST31AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS0": &Mount{Name: "TEST32AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS0": &Mount{Name: "TEST33AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS0": &Mount{Name: "TEST34AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS0": &Mount{Name: "TEST35AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS0": &Mount{Name: "TEST36AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS0": &Mount{Name: "TEST37AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS0": &Mount{Name: "TEST38AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS0": &Mount{Name: "TEST39AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS0": &Mount{Name: "TEST40AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS0": &Mount{Name: "TEST41AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS0": &Mount{Name: "TEST42AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS0": &Mount{Name: "TEST43AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS0": &Mount{Name: "TEST44AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS0": &Mount{Name: "TEST45AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS0": &Mount{Name: "TEST46AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS0": &Mount{Name: "TEST47AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS0": &Mount{Name: "TEST48AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS0": &Mount{Name: "TEST49AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS0": &Mount{Name: "TEST50AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS0": &Mount{Name: "TEST51AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS0": &Mount{Name: "TEST52AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS0": &Mount{Name: "TEST53AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS0": &Mount{Name: "TEST54AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS0": &Mount{Name: "TEST55AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS0": &Mount{Name: "TEST56AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS0": &Mount{Name: "TEST57AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS0": &Mount{Name: "TEST58AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS0": &Mount{Name: "TEST59AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS0": &Mount{Name: "TEST60AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS0": &Mount{Name: "TEST61AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS0": &Mount{Name: "TEST62AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS0": &Mount{Name: "TEST63AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS0": &Mount{Name: "TEST64AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS0": &Mount{Name: "TEST65AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS0": &Mount{Name: "TEST66AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS0": &Mount{Name: "TEST67AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS0": &Mount{Name: "TEST68AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS0": &Mount{Name: "TEST69AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS0": &Mount{Name: "TEST70AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS0": &Mount{Name: "TEST71AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS0": &Mount{Name: "TEST72AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS0": &Mount{Name: "TEST73AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS0": &Mount{Name: "TEST74AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS0": &Mount{Name: "TEST75AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS0": &Mount{Name: "TEST76AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS0": &Mount{Name: "TEST77AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS0": &Mount{Name: "TEST78AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS0": &Mount{Name: "TEST79AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS0": &Mount{Name: "TEST80AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS0": &Mount{Name: "TEST81AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS0": &Mount{Name: "TEST82AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS0": &Mount{Name: "TEST83AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS0": &Mount{Name: "TEST84AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS0": &Mount{Name: "TEST85AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS0": &Mount{Name: "TEST86AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS0": &Mount{Name: "TEST87AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS0": &Mount{Name: "TEST88AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS0": &Mount{Name: "TEST89AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS0": &Mount{Name: "TEST90AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS0": &Mount{Name: "TEST91AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS0": &Mount{Name: "TEST92AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS0": &Mount{Name: "TEST93AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS0": &Mount{Name: "TEST94AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS0": &Mount{Name: "TEST95AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS0": &Mount{Name: "TEST96AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS0": &Mount{Name: "TEST97AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS0": &Mount{Name: "TEST98AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS0": &Mount{Name: "TEST99AUS0", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS1": &Mount{Name: "TEST00AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS1": &Mount{Name: "TEST01AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS1": &Mount{Name: "TEST02AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS1": &Mount{Name: "TEST03AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS1": &Mount{Name: "TEST04AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS1": &Mount{Name: "TEST05AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS1": &Mount{Name: "TEST06AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS1": &Mount{Name: "TEST07AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS1": &Mount{Name: "TEST08AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS1": &Mount{Name: "TEST09AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS1": &Mount{Name: "TEST10AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS1": &Mount{Name: "TEST11AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS1": &Mount{Name: "TEST12AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS1": &Mount{Name: "TEST13AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS1": &Mount{Name: "TEST14AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS1": &Mount{Name: "TEST15AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS1": &Mount{Name: "TEST16AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS1": &Mount{Name: "TEST17AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS1": &Mount{Name: "TEST18AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS1": &Mount{Name: "TEST19AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS1": &Mount{Name: "TEST20AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS1": &Mount{Name: "TEST21AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS1": &Mount{Name: "TEST22AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS1": &Mount{Name: "TEST23AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS1": &Mount{Name: "TEST24AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS1": &Mount{Name: "TEST25AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS1": &Mount{Name: "TEST26AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS1": &Mount{Name: "TEST27AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS1": &Mount{Name: "TEST28AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS1": &Mount{Name: "TEST29AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS1": &Mount{Name: "TEST30AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS1": &Mount{Name: "TEST31AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS1": &Mount{Name: "TEST32AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS1": &Mount{Name: "TEST33AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS1": &Mount{Name: "TEST34AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS1": &Mount{Name: "TEST35AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS1": &Mount{Name: "TEST36AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS1": &Mount{Name: "TEST37AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS1": &Mount{Name: "TEST38AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS1": &Mount{Name: "TEST39AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS1": &Mount{Name: "TEST40AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS1": &Mount{Name: "TEST41AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS1": &Mount{Name: "TEST42AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS1": &Mount{Name: "TEST43AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS1": &Mount{Name: "TEST44AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS1": &Mount{Name: "TEST45AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS1": &Mount{Name: "TEST46AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS1": &Mount{Name: "TEST47AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS1": &Mount{Name: "TEST48AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS1": &Mount{Name: "TEST49AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS1": &Mount{Name: "TEST50AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS1": &Mount{Name: "TEST51AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS1": &Mount{Name: "TEST52AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS1": &Mount{Name: "TEST53AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS1": &Mount{Name: "TEST54AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS1": &Mount{Name: "TEST55AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS1": &Mount{Name: "TEST56AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS1": &Mount{Name: "TEST57AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS1": &Mount{Name: "TEST58AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS1": &Mount{Name: "TEST59AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS1": &Mount{Name: "TEST60AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS1": &Mount{Name: "TEST61AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS1": &Mount{Name: "TEST62AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS1": &Mount{Name: "TEST63AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS1": &Mount{Name: "TEST64AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS1": &Mount{Name: "TEST65AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS1": &Mount{Name: "TEST66AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS1": &Mount{Name: "TEST67AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS1": &Mount{Name: "TEST68AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS1": &Mount{Name: "TEST69AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS1": &Mount{Name: "TEST70AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS1": &Mount{Name: "TEST71AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS1": &Mount{Name: "TEST72AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS1": &Mount{Name: "TEST73AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS1": &Mount{Name: "TEST74AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS1": &Mount{Name: "TEST75AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS1": &Mount{Name: "TEST76AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS1": &Mount{Name: "TEST77AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS1": &Mount{Name: "TEST78AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS1": &Mount{Name: "TEST79AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS1": &Mount{Name: "TEST80AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS1": &Mount{Name: "TEST81AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS1": &Mount{Name: "TEST82AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS1": &Mount{Name: "TEST83AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS1": &Mount{Name: "TEST84AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS1": &Mount{Name: "TEST85AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS1": &Mount{Name: "TEST86AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS1": &Mount{Name: "TEST87AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS1": &Mount{Name: "TEST88AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS1": &Mount{Name: "TEST89AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS1": &Mount{Name: "TEST90AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS1": &Mount{Name: "TEST91AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS1": &Mount{Name: "TEST92AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS1": &Mount{Name: "TEST93AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS1": &Mount{Name: "TEST94AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS1": &Mount{Name: "TEST95AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS1": &Mount{Name: "TEST96AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS1": &Mount{Name: "TEST97AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS1": &Mount{Name: "TEST98AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS1": &Mount{Name: "TEST99AUS1", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS2": &Mount{Name: "TEST00AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS2": &Mount{Name: "TEST01AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS2": &Mount{Name: "TEST02AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS2": &Mount{Name: "TEST03AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS2": &Mount{Name: "TEST04AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS2": &Mount{Name: "TEST05AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS2": &Mount{Name: "TEST06AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS2": &Mount{Name: "TEST07AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS2": &Mount{Name: "TEST08AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS2": &Mount{Name: "TEST09AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS2": &Mount{Name: "TEST10AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS2": &Mount{Name: "TEST11AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS2": &Mount{Name: "TEST12AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS2": &Mount{Name: "TEST13AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS2": &Mount{Name: "TEST14AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS2": &Mount{Name: "TEST15AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS2": &Mount{Name: "TEST16AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS2": &Mount{Name: "TEST17AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS2": &Mount{Name: "TEST18AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS2": &Mount{Name: "TEST19AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS2": &Mount{Name: "TEST20AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS2": &Mount{Name: "TEST21AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS2": &Mount{Name: "TEST22AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS2": &Mount{Name: "TEST23AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS2": &Mount{Name: "TEST24AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS2": &Mount{Name: "TEST25AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS2": &Mount{Name: "TEST26AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS2": &Mount{Name: "TEST27AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS2": &Mount{Name: "TEST28AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS2": &Mount{Name: "TEST29AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS2": &Mount{Name: "TEST30AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS2": &Mount{Name: "TEST31AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS2": &Mount{Name: "TEST32AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS2": &Mount{Name: "TEST33AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS2": &Mount{Name: "TEST34AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS2": &Mount{Name: "TEST35AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS2": &Mount{Name: "TEST36AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS2": &Mount{Name: "TEST37AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS2": &Mount{Name: "TEST38AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS2": &Mount{Name: "TEST39AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS2": &Mount{Name: "TEST40AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS2": &Mount{Name: "TEST41AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS2": &Mount{Name: "TEST42AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS2": &Mount{Name: "TEST43AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS2": &Mount{Name: "TEST44AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS2": &Mount{Name: "TEST45AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS2": &Mount{Name: "TEST46AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS2": &Mount{Name: "TEST47AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS2": &Mount{Name: "TEST48AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS2": &Mount{Name: "TEST49AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS2": &Mount{Name: "TEST50AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS2": &Mount{Name: "TEST51AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS2": &Mount{Name: "TEST52AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS2": &Mount{Name: "TEST53AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS2": &Mount{Name: "TEST54AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS2": &Mount{Name: "TEST55AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS2": &Mount{Name: "TEST56AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS2": &Mount{Name: "TEST57AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS2": &Mount{Name: "TEST58AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS2": &Mount{Name: "TEST59AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS2": &Mount{Name: "TEST60AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS2": &Mount{Name: "TEST61AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS2": &Mount{Name: "TEST62AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS2": &Mount{Name: "TEST63AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS2": &Mount{Name: "TEST64AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS2": &Mount{Name: "TEST65AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS2": &Mount{Name: "TEST66AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS2": &Mount{Name: "TEST67AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS2": &Mount{Name: "TEST68AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS2": &Mount{Name: "TEST69AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS2": &Mount{Name: "TEST70AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS2": &Mount{Name: "TEST71AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS2": &Mount{Name: "TEST72AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS2": &Mount{Name: "TEST73AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS2": &Mount{Name: "TEST74AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS2": &Mount{Name: "TEST75AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS2": &Mount{Name: "TEST76AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS2": &Mount{Name: "TEST77AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS2": &Mount{Name: "TEST78AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS2": &Mount{Name: "TEST79AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS2": &Mount{Name: "TEST80AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS2": &Mount{Name: "TEST81AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS2": &Mount{Name: "TEST82AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS2": &Mount{Name: "TEST83AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS2": &Mount{Name: "TEST84AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS2": &Mount{Name: "TEST85AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS2": &Mount{Name: "TEST86AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS2": &Mount{Name: "TEST87AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS2": &Mount{Name: "TEST88AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS2": &Mount{Name: "TEST89AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS2": &Mount{Name: "TEST90AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS2": &Mount{Name: "TEST91AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS2": &Mount{Name: "TEST92AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS2": &Mount{Name: "TEST93AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS2": &Mount{Name: "TEST94AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS2": &Mount{Name: "TEST95AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS2": &Mount{Name: "TEST96AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS2": &Mount{Name: "TEST97AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS2": &Mount{Name: "TEST98AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS2": &Mount{Name: "TEST99AUS2", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS3": &Mount{Name: "TEST00AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS3": &Mount{Name: "TEST01AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS3": &Mount{Name: "TEST02AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS3": &Mount{Name: "TEST03AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS3": &Mount{Name: "TEST04AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS3": &Mount{Name: "TEST05AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS3": &Mount{Name: "TEST06AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS3": &Mount{Name: "TEST07AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS3": &Mount{Name: "TEST08AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS3": &Mount{Name: "TEST09AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS3": &Mount{Name: "TEST10AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS3": &Mount{Name: "TEST11AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS3": &Mount{Name: "TEST12AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS3": &Mount{Name: "TEST13AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS3": &Mount{Name: "TEST14AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS3": &Mount{Name: "TEST15AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS3": &Mount{Name: "TEST16AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS3": &Mount{Name: "TEST17AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS3": &Mount{Name: "TEST18AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS3": &Mount{Name: "TEST19AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS3": &Mount{Name: "TEST20AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS3": &Mount{Name: "TEST21AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS3": &Mount{Name: "TEST22AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS3": &Mount{Name: "TEST23AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS3": &Mount{Name: "TEST24AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS3": &Mount{Name: "TEST25AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS3": &Mount{Name: "TEST26AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS3": &Mount{Name: "TEST27AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS3": &Mount{Name: "TEST28AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS3": &Mount{Name: "TEST29AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS3": &Mount{Name: "TEST30AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS3": &Mount{Name: "TEST31AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS3": &Mount{Name: "TEST32AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS3": &Mount{Name: "TEST33AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS3": &Mount{Name: "TEST34AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS3": &Mount{Name: "TEST35AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS3": &Mount{Name: "TEST36AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS3": &Mount{Name: "TEST37AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS3": &Mount{Name: "TEST38AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS3": &Mount{Name: "TEST39AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS3": &Mount{Name: "TEST40AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS3": &Mount{Name: "TEST41AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS3": &Mount{Name: "TEST42AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS3": &Mount{Name: "TEST43AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS3": &Mount{Name: "TEST44AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS3": &Mount{Name: "TEST45AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS3": &Mount{Name: "TEST46AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS3": &Mount{Name: "TEST47AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS3": &Mount{Name: "TEST48AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS3": &Mount{Name: "TEST49AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS3": &Mount{Name: "TEST50AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS3": &Mount{Name: "TEST51AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS3": &Mount{Name: "TEST52AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS3": &Mount{Name: "TEST53AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS3": &Mount{Name: "TEST54AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS3": &Mount{Name: "TEST55AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS3": &Mount{Name: "TEST56AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS3": &Mount{Name: "TEST57AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS3": &Mount{Name: "TEST58AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS3": &Mount{Name: "TEST59AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS3": &Mount{Name: "TEST60AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS3": &Mount{Name: "TEST61AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS3": &Mount{Name: "TEST62AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS3": &Mount{Name: "TEST63AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS3": &Mount{Name: "TEST64AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS3": &Mount{Name: "TEST65AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS3": &Mount{Name: "TEST66AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS3": &Mount{Name: "TEST67AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS3": &Mount{Name: "TEST68AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS3": &Mount{Name: "TEST69AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS3": &Mount{Name: "TEST70AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS3": &Mount{Name: "TEST71AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS3": &Mount{Name: "TEST72AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS3": &Mount{Name: "TEST73AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS3": &Mount{Name: "TEST74AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS3": &Mount{Name: "TEST75AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS3": &Mount{Name: "TEST76AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS3": &Mount{Name: "TEST77AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS3": &Mount{Name: "TEST78AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS3": &Mount{Name: "TEST79AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS3": &Mount{Name: "TEST80AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS3": &Mount{Name: "TEST81AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS3": &Mount{Name: "TEST82AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS3": &Mount{Name: "TEST83AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS3": &Mount{Name: "TEST84AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS3": &Mount{Name: "TEST85AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS3": &Mount{Name: "TEST86AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS3": &Mount{Name: "TEST87AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS3": &Mount{Name: "TEST88AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS3": &Mount{Name: "TEST89AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS3": &Mount{Name: "TEST90AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS3": &Mount{Name: "TEST91AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS3": &Mount{Name: "TEST92AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS3": &Mount{Name: "TEST93AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS3": &Mount{Name: "TEST94AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS3": &Mount{Name: "TEST95AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS3": &Mount{Name: "TEST96AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS3": &Mount{Name: "TEST97AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS3": &Mount{Name: "TEST98AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS3": &Mount{Name: "TEST99AUS3", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS4": &Mount{Name: "TEST00AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS4": &Mount{Name: "TEST01AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS4": &Mount{Name: "TEST02AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS4": &Mount{Name: "TEST03AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS4": &Mount{Name: "TEST04AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS4": &Mount{Name: "TEST05AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS4": &Mount{Name: "TEST06AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS4": &Mount{Name: "TEST07AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS4": &Mount{Name: "TEST08AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS4": &Mount{Name: "TEST09AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS4": &Mount{Name: "TEST10AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS4": &Mount{Name: "TEST11AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS4": &Mount{Name: "TEST12AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS4": &Mount{Name: "TEST13AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS4": &Mount{Name: "TEST14AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS4": &Mount{Name: "TEST15AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS4": &Mount{Name: "TEST16AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS4": &Mount{Name: "TEST17AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS4": &Mount{Name: "TEST18AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS4": &Mount{Name: "TEST19AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS4": &Mount{Name: "TEST20AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS4": &Mount{Name: "TEST21AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS4": &Mount{Name: "TEST22AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS4": &Mount{Name: "TEST23AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS4": &Mount{Name: "TEST24AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS4": &Mount{Name: "TEST25AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS4": &Mount{Name: "TEST26AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS4": &Mount{Name: "TEST27AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS4": &Mount{Name: "TEST28AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS4": &Mount{Name: "TEST29AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS4": &Mount{Name: "TEST30AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS4": &Mount{Name: "TEST31AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS4": &Mount{Name: "TEST32AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS4": &Mount{Name: "TEST33AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS4": &Mount{Name: "TEST34AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS4": &Mount{Name: "TEST35AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS4": &Mount{Name: "TEST36AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS4": &Mount{Name: "TEST37AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS4": &Mount{Name: "TEST38AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS4": &Mount{Name: "TEST39AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS4": &Mount{Name: "TEST40AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS4": &Mount{Name: "TEST41AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS4": &Mount{Name: "TEST42AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS4": &Mount{Name: "TEST43AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS4": &Mount{Name: "TEST44AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS4": &Mount{Name: "TEST45AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS4": &Mount{Name: "TEST46AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS4": &Mount{Name: "TEST47AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS4": &Mount{Name: "TEST48AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS4": &Mount{Name: "TEST49AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS4": &Mount{Name: "TEST50AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS4": &Mount{Name: "TEST51AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS4": &Mount{Name: "TEST52AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS4": &Mount{Name: "TEST53AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS4": &Mount{Name: "TEST54AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS4": &Mount{Name: "TEST55AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS4": &Mount{Name: "TEST56AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS4": &Mount{Name: "TEST57AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS4": &Mount{Name: "TEST58AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS4": &Mount{Name: "TEST59AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS4": &Mount{Name: "TEST60AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS4": &Mount{Name: "TEST61AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS4": &Mount{Name: "TEST62AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS4": &Mount{Name: "TEST63AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS4": &Mount{Name: "TEST64AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS4": &Mount{Name: "TEST65AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS4": &Mount{Name: "TEST66AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS4": &Mount{Name: "TEST67AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS4": &Mount{Name: "TEST68AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS4": &Mount{Name: "TEST69AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS4": &Mount{Name: "TEST70AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS4": &Mount{Name: "TEST71AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS4": &Mount{Name: "TEST72AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS4": &Mount{Name: "TEST73AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS4": &Mount{Name: "TEST74AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS4": &Mount{Name: "TEST75AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS4": &Mount{Name: "TEST76AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS4": &Mount{Name: "TEST77AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS4": &Mount{Name: "TEST78AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS4": &Mount{Name: "TEST79AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS4": &Mount{Name: "TEST80AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS4": &Mount{Name: "TEST81AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS4": &Mount{Name: "TEST82AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS4": &Mount{Name: "TEST83AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS4": &Mount{Name: "TEST84AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS4": &Mount{Name: "TEST85AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS4": &Mount{Name: "TEST86AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS4": &Mount{Name: "TEST87AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS4": &Mount{Name: "TEST88AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS4": &Mount{Name: "TEST89AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS4": &Mount{Name: "TEST90AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS4": &Mount{Name: "TEST91AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS4": &Mount{Name: "TEST92AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS4": &Mount{Name: "TEST93AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS4": &Mount{Name: "TEST94AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS4": &Mount{Name: "TEST95AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS4": &Mount{Name: "TEST96AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS4": &Mount{Name: "TEST97AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS4": &Mount{Name: "TEST98AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS4": &Mount{Name: "TEST99AUS4", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS5": &Mount{Name: "TEST00AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS5": &Mount{Name: "TEST01AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS5": &Mount{Name: "TEST02AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS5": &Mount{Name: "TEST03AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS5": &Mount{Name: "TEST04AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS5": &Mount{Name: "TEST05AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS5": &Mount{Name: "TEST06AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS5": &Mount{Name: "TEST07AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS5": &Mount{Name: "TEST08AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS5": &Mount{Name: "TEST09AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS5": &Mount{Name: "TEST10AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS5": &Mount{Name: "TEST11AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS5": &Mount{Name: "TEST12AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS5": &Mount{Name: "TEST13AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS5": &Mount{Name: "TEST14AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS5": &Mount{Name: "TEST15AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS5": &Mount{Name: "TEST16AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS5": &Mount{Name: "TEST17AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS5": &Mount{Name: "TEST18AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS5": &Mount{Name: "TEST19AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS5": &Mount{Name: "TEST20AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS5": &Mount{Name: "TEST21AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS5": &Mount{Name: "TEST22AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS5": &Mount{Name: "TEST23AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS5": &Mount{Name: "TEST24AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS5": &Mount{Name: "TEST25AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS5": &Mount{Name: "TEST26AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS5": &Mount{Name: "TEST27AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS5": &Mount{Name: "TEST28AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS5": &Mount{Name: "TEST29AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS5": &Mount{Name: "TEST30AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS5": &Mount{Name: "TEST31AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS5": &Mount{Name: "TEST32AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS5": &Mount{Name: "TEST33AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS5": &Mount{Name: "TEST34AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS5": &Mount{Name: "TEST35AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS5": &Mount{Name: "TEST36AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS5": &Mount{Name: "TEST37AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS5": &Mount{Name: "TEST38AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS5": &Mount{Name: "TEST39AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS5": &Mount{Name: "TEST40AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS5": &Mount{Name: "TEST41AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS5": &Mount{Name: "TEST42AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS5": &Mount{Name: "TEST43AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS5": &Mount{Name: "TEST44AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS5": &Mount{Name: "TEST45AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS5": &Mount{Name: "TEST46AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS5": &Mount{Name: "TEST47AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS5": &Mount{Name: "TEST48AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS5": &Mount{Name: "TEST49AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS5": &Mount{Name: "TEST50AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS5": &Mount{Name: "TEST51AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS5": &Mount{Name: "TEST52AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS5": &Mount{Name: "TEST53AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS5": &Mount{Name: "TEST54AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS5": &Mount{Name: "TEST55AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS5": &Mount{Name: "TEST56AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS5": &Mount{Name: "TEST57AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS5": &Mount{Name: "TEST58AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS5": &Mount{Name: "TEST59AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS5": &Mount{Name: "TEST60AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS5": &Mount{Name: "TEST61AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS5": &Mount{Name: "TEST62AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS5": &Mount{Name: "TEST63AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS5": &Mount{Name: "TEST64AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS5": &Mount{Name: "TEST65AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS5": &Mount{Name: "TEST66AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS5": &Mount{Name: "TEST67AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS5": &Mount{Name: "TEST68AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS5": &Mount{Name: "TEST69AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS5": &Mount{Name: "TEST70AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS5": &Mount{Name: "TEST71AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS5": &Mount{Name: "TEST72AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS5": &Mount{Name: "TEST73AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS5": &Mount{Name: "TEST74AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS5": &Mount{Name: "TEST75AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS5": &Mount{Name: "TEST76AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS5": &Mount{Name: "TEST77AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS5": &Mount{Name: "TEST78AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS5": &Mount{Name: "TEST79AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS5": &Mount{Name: "TEST80AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS5": &Mount{Name: "TEST81AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS5": &Mount{Name: "TEST82AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS5": &Mount{Name: "TEST83AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS5": &Mount{Name: "TEST84AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS5": &Mount{Name: "TEST85AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS5": &Mount{Name: "TEST86AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS5": &Mount{Name: "TEST87AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS5": &Mount{Name: "TEST88AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS5": &Mount{Name: "TEST89AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS5": &Mount{Name: "TEST90AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS5": &Mount{Name: "TEST91AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS5": &Mount{Name: "TEST92AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS5": &Mount{Name: "TEST93AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS5": &Mount{Name: "TEST94AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS5": &Mount{Name: "TEST95AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS5": &Mount{Name: "TEST96AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS5": &Mount{Name: "TEST97AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS5": &Mount{Name: "TEST98AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS5": &Mount{Name: "TEST99AUS5", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS6": &Mount{Name: "TEST00AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS6": &Mount{Name: "TEST01AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS6": &Mount{Name: "TEST02AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS6": &Mount{Name: "TEST03AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS6": &Mount{Name: "TEST04AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS6": &Mount{Name: "TEST05AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS6": &Mount{Name: "TEST06AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS6": &Mount{Name: "TEST07AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS6": &Mount{Name: "TEST08AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS6": &Mount{Name: "TEST09AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS6": &Mount{Name: "TEST10AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS6": &Mount{Name: "TEST11AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS6": &Mount{Name: "TEST12AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS6": &Mount{Name: "TEST13AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS6": &Mount{Name: "TEST14AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS6": &Mount{Name: "TEST15AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS6": &Mount{Name: "TEST16AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS6": &Mount{Name: "TEST17AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS6": &Mount{Name: "TEST18AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS6": &Mount{Name: "TEST19AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS6": &Mount{Name: "TEST20AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS6": &Mount{Name: "TEST21AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS6": &Mount{Name: "TEST22AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS6": &Mount{Name: "TEST23AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS6": &Mount{Name: "TEST24AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS6": &Mount{Name: "TEST25AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS6": &Mount{Name: "TEST26AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS6": &Mount{Name: "TEST27AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS6": &Mount{Name: "TEST28AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS6": &Mount{Name: "TEST29AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS6": &Mount{Name: "TEST30AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS6": &Mount{Name: "TEST31AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS6": &Mount{Name: "TEST32AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS6": &Mount{Name: "TEST33AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS6": &Mount{Name: "TEST34AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS6": &Mount{Name: "TEST35AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS6": &Mount{Name: "TEST36AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS6": &Mount{Name: "TEST37AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS6": &Mount{Name: "TEST38AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS6": &Mount{Name: "TEST39AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS6": &Mount{Name: "TEST40AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS6": &Mount{Name: "TEST41AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS6": &Mount{Name: "TEST42AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS6": &Mount{Name: "TEST43AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS6": &Mount{Name: "TEST44AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS6": &Mount{Name: "TEST45AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS6": &Mount{Name: "TEST46AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS6": &Mount{Name: "TEST47AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS6": &Mount{Name: "TEST48AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS6": &Mount{Name: "TEST49AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS6": &Mount{Name: "TEST50AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS6": &Mount{Name: "TEST51AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS6": &Mount{Name: "TEST52AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS6": &Mount{Name: "TEST53AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS6": &Mount{Name: "TEST54AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS6": &Mount{Name: "TEST55AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS6": &Mount{Name: "TEST56AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS6": &Mount{Name: "TEST57AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS6": &Mount{Name: "TEST58AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS6": &Mount{Name: "TEST59AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS6": &Mount{Name: "TEST60AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS6": &Mount{Name: "TEST61AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS6": &Mount{Name: "TEST62AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS6": &Mount{Name: "TEST63AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS6": &Mount{Name: "TEST64AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS6": &Mount{Name: "TEST65AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS6": &Mount{Name: "TEST66AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS6": &Mount{Name: "TEST67AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS6": &Mount{Name: "TEST68AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS6": &Mount{Name: "TEST69AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS6": &Mount{Name: "TEST70AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS6": &Mount{Name: "TEST71AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS6": &Mount{Name: "TEST72AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS6": &Mount{Name: "TEST73AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS6": &Mount{Name: "TEST74AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS6": &Mount{Name: "TEST75AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS6": &Mount{Name: "TEST76AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS6": &Mount{Name: "TEST77AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS6": &Mount{Name: "TEST78AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS6": &Mount{Name: "TEST79AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS6": &Mount{Name: "TEST80AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS6": &Mount{Name: "TEST81AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS6": &Mount{Name: "TEST82AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS6": &Mount{Name: "TEST83AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS6": &Mount{Name: "TEST84AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS6": &Mount{Name: "TEST85AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS6": &Mount{Name: "TEST86AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS6": &Mount{Name: "TEST87AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS6": &Mount{Name: "TEST88AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS6": &Mount{Name: "TEST89AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS6": &Mount{Name: "TEST90AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS6": &Mount{Name: "TEST91AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS6": &Mount{Name: "TEST92AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS6": &Mount{Name: "TEST93AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS6": &Mount{Name: "TEST94AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS6": &Mount{Name: "TEST95AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS6": &Mount{Name: "TEST96AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS6": &Mount{Name: "TEST97AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS6": &Mount{Name: "TEST98AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS6": &Mount{Name: "TEST99AUS6", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS7": &Mount{Name: "TEST00AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS7": &Mount{Name: "TEST01AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS7": &Mount{Name: "TEST02AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS7": &Mount{Name: "TEST03AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS7": &Mount{Name: "TEST04AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS7": &Mount{Name: "TEST05AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS7": &Mount{Name: "TEST06AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS7": &Mount{Name: "TEST07AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS7": &Mount{Name: "TEST08AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS7": &Mount{Name: "TEST09AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS7": &Mount{Name: "TEST10AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS7": &Mount{Name: "TEST11AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS7": &Mount{Name: "TEST12AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS7": &Mount{Name: "TEST13AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS7": &Mount{Name: "TEST14AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS7": &Mount{Name: "TEST15AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS7": &Mount{Name: "TEST16AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS7": &Mount{Name: "TEST17AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS7": &Mount{Name: "TEST18AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS7": &Mount{Name: "TEST19AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS7": &Mount{Name: "TEST20AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS7": &Mount{Name: "TEST21AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS7": &Mount{Name: "TEST22AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS7": &Mount{Name: "TEST23AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS7": &Mount{Name: "TEST24AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS7": &Mount{Name: "TEST25AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS7": &Mount{Name: "TEST26AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS7": &Mount{Name: "TEST27AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS7": &Mount{Name: "TEST28AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS7": &Mount{Name: "TEST29AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS7": &Mount{Name: "TEST30AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS7": &Mount{Name: "TEST31AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS7": &Mount{Name: "TEST32AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS7": &Mount{Name: "TEST33AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS7": &Mount{Name: "TEST34AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS7": &Mount{Name: "TEST35AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS7": &Mount{Name: "TEST36AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS7": &Mount{Name: "TEST37AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS7": &Mount{Name: "TEST38AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS7": &Mount{Name: "TEST39AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS7": &Mount{Name: "TEST40AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS7": &Mount{Name: "TEST41AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS7": &Mount{Name: "TEST42AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS7": &Mount{Name: "TEST43AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS7": &Mount{Name: "TEST44AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS7": &Mount{Name: "TEST45AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS7": &Mount{Name: "TEST46AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS7": &Mount{Name: "TEST47AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS7": &Mount{Name: "TEST48AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS7": &Mount{Name: "TEST49AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS7": &Mount{Name: "TEST50AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS7": &Mount{Name: "TEST51AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS7": &Mount{Name: "TEST52AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS7": &Mount{Name: "TEST53AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS7": &Mount{Name: "TEST54AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS7": &Mount{Name: "TEST55AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS7": &Mount{Name: "TEST56AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS7": &Mount{Name: "TEST57AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS7": &Mount{Name: "TEST58AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS7": &Mount{Name: "TEST59AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS7": &Mount{Name: "TEST60AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS7": &Mount{Name: "TEST61AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS7": &Mount{Name: "TEST62AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS7": &Mount{Name: "TEST63AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS7": &Mount{Name: "TEST64AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS7": &Mount{Name: "TEST65AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS7": &Mount{Name: "TEST66AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS7": &Mount{Name: "TEST67AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS7": &Mount{Name: "TEST68AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS7": &Mount{Name: "TEST69AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS7": &Mount{Name: "TEST70AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS7": &Mount{Name: "TEST71AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS7": &Mount{Name: "TEST72AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS7": &Mount{Name: "TEST73AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS7": &Mount{Name: "TEST74AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS7": &Mount{Name: "TEST75AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS7": &Mount{Name: "TEST76AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS7": &Mount{Name: "TEST77AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS7": &Mount{Name: "TEST78AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS7": &Mount{Name: "TEST79AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS7": &Mount{Name: "TEST80AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS7": &Mount{Name: "TEST81AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS7": &Mount{Name: "TEST82AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS7": &Mount{Name: "TEST83AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS7": &Mount{Name: "TEST84AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS7": &Mount{Name: "TEST85AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS7": &Mount{Name: "TEST86AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS7": &Mount{Name: "TEST87AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS7": &Mount{Name: "TEST88AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS7": &Mount{Name: "TEST89AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS7": &Mount{Name: "TEST90AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS7": &Mount{Name: "TEST91AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS7": &Mount{Name: "TEST92AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS7": &Mount{Name: "TEST93AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS7": &Mount{Name: "TEST94AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS7": &Mount{Name: "TEST95AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS7": &Mount{Name: "TEST96AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS7": &Mount{Name: "TEST97AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS7": &Mount{Name: "TEST98AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS7": &Mount{Name: "TEST99AUS7", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS8": &Mount{Name: "TEST00AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS8": &Mount{Name: "TEST01AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS8": &Mount{Name: "TEST02AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS8": &Mount{Name: "TEST03AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS8": &Mount{Name: "TEST04AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS8": &Mount{Name: "TEST05AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS8": &Mount{Name: "TEST06AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS8": &Mount{Name: "TEST07AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS8": &Mount{Name: "TEST08AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS8": &Mount{Name: "TEST09AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS8": &Mount{Name: "TEST10AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS8": &Mount{Name: "TEST11AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS8": &Mount{Name: "TEST12AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS8": &Mount{Name: "TEST13AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS8": &Mount{Name: "TEST14AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS8": &Mount{Name: "TEST15AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS8": &Mount{Name: "TEST16AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS8": &Mount{Name: "TEST17AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS8": &Mount{Name: "TEST18AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS8": &Mount{Name: "TEST19AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS8": &Mount{Name: "TEST20AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS8": &Mount{Name: "TEST21AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS8": &Mount{Name: "TEST22AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS8": &Mount{Name: "TEST23AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS8": &Mount{Name: "TEST24AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS8": &Mount{Name: "TEST25AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS8": &Mount{Name: "TEST26AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS8": &Mount{Name: "TEST27AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS8": &Mount{Name: "TEST28AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS8": &Mount{Name: "TEST29AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS8": &Mount{Name: "TEST30AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS8": &Mount{Name: "TEST31AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS8": &Mount{Name: "TEST32AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS8": &Mount{Name: "TEST33AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS8": &Mount{Name: "TEST34AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS8": &Mount{Name: "TEST35AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS8": &Mount{Name: "TEST36AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS8": &Mount{Name: "TEST37AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS8": &Mount{Name: "TEST38AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS8": &Mount{Name: "TEST39AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS8": &Mount{Name: "TEST40AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS8": &Mount{Name: "TEST41AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS8": &Mount{Name: "TEST42AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS8": &Mount{Name: "TEST43AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS8": &Mount{Name: "TEST44AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS8": &Mount{Name: "TEST45AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS8": &Mount{Name: "TEST46AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS8": &Mount{Name: "TEST47AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS8": &Mount{Name: "TEST48AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS8": &Mount{Name: "TEST49AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS8": &Mount{Name: "TEST50AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS8": &Mount{Name: "TEST51AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS8": &Mount{Name: "TEST52AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS8": &Mount{Name: "TEST53AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS8": &Mount{Name: "TEST54AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS8": &Mount{Name: "TEST55AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS8": &Mount{Name: "TEST56AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS8": &Mount{Name: "TEST57AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS8": &Mount{Name: "TEST58AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS8": &Mount{Name: "TEST59AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS8": &Mount{Name: "TEST60AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS8": &Mount{Name: "TEST61AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS8": &Mount{Name: "TEST62AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS8": &Mount{Name: "TEST63AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS8": &Mount{Name: "TEST64AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS8": &Mount{Name: "TEST65AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS8": &Mount{Name: "TEST66AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS8": &Mount{Name: "TEST67AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS8": &Mount{Name: "TEST68AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS8": &Mount{Name: "TEST69AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS8": &Mount{Name: "TEST70AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS8": &Mount{Name: "TEST71AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS8": &Mount{Name: "TEST72AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS8": &Mount{Name: "TEST73AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS8": &Mount{Name: "TEST74AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS8": &Mount{Name: "TEST75AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS8": &Mount{Name: "TEST76AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS8": &Mount{Name: "TEST77AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS8": &Mount{Name: "TEST78AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS8": &Mount{Name: "TEST79AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS8": &Mount{Name: "TEST80AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS8": &Mount{Name: "TEST81AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS8": &Mount{Name: "TEST82AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS8": &Mount{Name: "TEST83AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS8": &Mount{Name: "TEST84AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS8": &Mount{Name: "TEST85AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS8": &Mount{Name: "TEST86AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS8": &Mount{Name: "TEST87AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS8": &Mount{Name: "TEST88AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS8": &Mount{Name: "TEST89AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS8": &Mount{Name: "TEST90AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS8": &Mount{Name: "TEST91AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS8": &Mount{Name: "TEST92AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS8": &Mount{Name: "TEST93AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS8": &Mount{Name: "TEST94AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS8": &Mount{Name: "TEST95AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS8": &Mount{Name: "TEST96AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS8": &Mount{Name: "TEST97AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS8": &Mount{Name: "TEST98AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS8": &Mount{Name: "TEST99AUS8", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST00AUS9": &Mount{Name: "TEST00AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST01AUS9": &Mount{Name: "TEST01AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST02AUS9": &Mount{Name: "TEST02AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST03AUS9": &Mount{Name: "TEST03AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST04AUS9": &Mount{Name: "TEST04AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST05AUS9": &Mount{Name: "TEST05AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST06AUS9": &Mount{Name: "TEST06AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST07AUS9": &Mount{Name: "TEST07AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST08AUS9": &Mount{Name: "TEST08AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST09AUS9": &Mount{Name: "TEST09AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST10AUS9": &Mount{Name: "TEST10AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST11AUS9": &Mount{Name: "TEST11AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST12AUS9": &Mount{Name: "TEST12AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST13AUS9": &Mount{Name: "TEST13AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST14AUS9": &Mount{Name: "TEST14AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST15AUS9": &Mount{Name: "TEST15AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST16AUS9": &Mount{Name: "TEST16AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST17AUS9": &Mount{Name: "TEST17AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST18AUS9": &Mount{Name: "TEST18AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST19AUS9": &Mount{Name: "TEST19AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST20AUS9": &Mount{Name: "TEST20AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST21AUS9": &Mount{Name: "TEST21AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST22AUS9": &Mount{Name: "TEST22AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST23AUS9": &Mount{Name: "TEST23AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST24AUS9": &Mount{Name: "TEST24AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST25AUS9": &Mount{Name: "TEST25AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST26AUS9": &Mount{Name: "TEST26AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST27AUS9": &Mount{Name: "TEST27AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST28AUS9": &Mount{Name: "TEST28AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST29AUS9": &Mount{Name: "TEST29AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST30AUS9": &Mount{Name: "TEST30AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST31AUS9": &Mount{Name: "TEST31AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST32AUS9": &Mount{Name: "TEST32AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST33AUS9": &Mount{Name: "TEST33AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST34AUS9": &Mount{Name: "TEST34AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST35AUS9": &Mount{Name: "TEST35AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST36AUS9": &Mount{Name: "TEST36AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST37AUS9": &Mount{Name: "TEST37AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST38AUS9": &Mount{Name: "TEST38AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST39AUS9": &Mount{Name: "TEST39AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST40AUS9": &Mount{Name: "TEST40AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST41AUS9": &Mount{Name: "TEST41AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST42AUS9": &Mount{Name: "TEST42AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST43AUS9": &Mount{Name: "TEST43AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST44AUS9": &Mount{Name: "TEST44AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST45AUS9": &Mount{Name: "TEST45AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST46AUS9": &Mount{Name: "TEST46AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST47AUS9": &Mount{Name: "TEST47AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST48AUS9": &Mount{Name: "TEST48AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST49AUS9": &Mount{Name: "TEST49AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST50AUS9": &Mount{Name: "TEST50AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST51AUS9": &Mount{Name: "TEST51AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST52AUS9": &Mount{Name: "TEST52AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST53AUS9": &Mount{Name: "TEST53AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST54AUS9": &Mount{Name: "TEST54AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST55AUS9": &Mount{Name: "TEST55AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST56AUS9": &Mount{Name: "TEST56AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST57AUS9": &Mount{Name: "TEST57AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST58AUS9": &Mount{Name: "TEST58AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST59AUS9": &Mount{Name: "TEST59AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST60AUS9": &Mount{Name: "TEST60AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST61AUS9": &Mount{Name: "TEST61AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST62AUS9": &Mount{Name: "TEST62AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST63AUS9": &Mount{Name: "TEST63AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST64AUS9": &Mount{Name: "TEST64AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST65AUS9": &Mount{Name: "TEST65AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST66AUS9": &Mount{Name: "TEST66AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST67AUS9": &Mount{Name: "TEST67AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST68AUS9": &Mount{Name: "TEST68AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST69AUS9": &Mount{Name: "TEST69AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST70AUS9": &Mount{Name: "TEST70AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST71AUS9": &Mount{Name: "TEST71AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST72AUS9": &Mount{Name: "TEST72AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST73AUS9": &Mount{Name: "TEST73AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST74AUS9": &Mount{Name: "TEST74AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST75AUS9": &Mount{Name: "TEST75AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST76AUS9": &Mount{Name: "TEST76AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST77AUS9": &Mount{Name: "TEST77AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST78AUS9": &Mount{Name: "TEST78AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST79AUS9": &Mount{Name: "TEST79AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST80AUS9": &Mount{Name: "TEST80AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST81AUS9": &Mount{Name: "TEST81AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST82AUS9": &Mount{Name: "TEST82AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST83AUS9": &Mount{Name: "TEST83AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST84AUS9": &Mount{Name: "TEST84AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST85AUS9": &Mount{Name: "TEST85AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST86AUS9": &Mount{Name: "TEST86AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST87AUS9": &Mount{Name: "TEST87AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST88AUS9": &Mount{Name: "TEST88AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST89AUS9": &Mount{Name: "TEST89AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST90AUS9": &Mount{Name: "TEST90AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST91AUS9": &Mount{Name: "TEST91AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST92AUS9": &Mount{Name: "TEST92AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST93AUS9": &Mount{Name: "TEST93AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST94AUS9": &Mount{Name: "TEST94AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST95AUS9": &Mount{Name: "TEST95AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST96AUS9": &Mount{Name: "TEST96AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST97AUS9": &Mount{Name: "TEST97AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST98AUS9": &Mount{Name: "TEST98AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-		"TEST99AUS9": &Mount{Name: "TEST99AUS9", Identifier: "Canberra (ACT)", Format: "RTCM 3.3"},
-	}, "tcp://vernemq:1883"}
-
-	log.SetFormatter(&log.JSONFormatter{})
-
-	// Watcher subscriptions update LastReceived attribute on Mount objects so 404s and timeouts can be implemented
-	// TODO: Create these subscriptions on initialization of / changes to config
-	mqttClient := mqtt.NewClient(mqtt.NewClientOptions().AddBroker(caster.Broker))
-	if token := mqttClient.Connect(); token.Wait() && token.Error() != nil {
-		log.Fatal("failed to create MQTT client - " + token.Error().Error())
-	}
-	defer mqttClient.Disconnect(100)
-
-	// There should be no harm in just resubscribing for all mounts on any change to config
-	for _, mount := range caster.Mounts {
-		func(mount *Mount) { // This doesn't need to be a function, but mount does need to be copied so the anonymous function passed to Subscribe isn't a closure referencing the for loop's mount variable
-			token := mqttClient.Subscribe(mount.Name + "/#", 1, func(_ mqtt.Client, msg mqtt.Message) {
-				mount.LastMessage = time.Now()
-			})
-			if token.Wait() && token.Error() != nil {
-				log.Fatal("MQTT subscription failed - " + token.Error().Error()) // Should this be fatal?
-			}
-		}(mount)
-	}
-
-	httpMux := mux.NewRouter()
-	httpMux.HandleFunc("/", caster.GetSourcetable).Methods("GET")
-	httpMux.HandleFunc("/{mountpoint}", caster.GetMount).Methods("GET")
-	httpMux.HandleFunc("/{mountpoint}", caster.PostMount).Methods("POST")
-
-	// This is probably fancier than it needs to be, could really just have a GetLogger function so we don't have to cast when pulling out the logger
-	// Perhaps a mix of both makes sense, where the UUID is added to context, but you generate a logger from the Request (which includes the context) when you need it
-	// One benefit of defining the logger in the Context is that it can be appended to
-	httpMux.Use(func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			requestId := uuid.New().String()
-			ctx := context.WithValue(r.Context(), "UUID", requestId)
-			ctx = context.WithValue(ctx, "logger", log.WithFields(log.Fields{
-				"request_id": requestId,
-				"path":       r.URL.Path,
-				"method":     r.Method,
-				"source_ip":  r.RemoteAddr,
-			}))
-			next.ServeHTTP(w, r.WithContext(ctx))
-		})
-	})
-
-	log.Info("server starting")
-	log.Fatal(http.ListenAndServe(":" + caster.Port, httpMux))
-}
